refactor(workspace): extract fractional ordering to shared/fractional-index
computeMidpoint and generateInitialOrders are generic ordering utilities
with no sheet/timeline dependency. Move them to shared/ so they can be
reused without pulling in the timeline module.
